Make CTRL+C handling consistent among tools
